requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
